# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Recipe Coster API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Ingredient {
  id: number;
  name: string;
  /** @nullable */
  supplier?: string | null;
  /** The unit the ingredient is purchased in (e.g. kg, L, each) */
  purchaseUnit: string;
  /** The size of the purchase unit (e.g. 1 for 1kg, 5 for 5L) */
  purchaseUnitSize: number;
  /** Cost of one purchase unit in dollars */
  purchaseCost: number;
  /** The unit used in recipes (e.g. g, ml, each) */
  recipeUnit: string;
  /** Cost per recipe unit (auto-calculated) */
  recipeUnitCost: number;
  /** @nullable */
  category?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateIngredientBody {
  name: string;
  /** @nullable */
  supplier?: string | null;
  purchaseUnit: string;
  purchaseUnitSize: number;
  purchaseCost: number;
  recipeUnit: string;
  /** @nullable */
  category?: string | null;
}

export interface UpdateIngredientBody {
  name?: string;
  /** @nullable */
  supplier?: string | null;
  purchaseUnit?: string;
  purchaseUnitSize?: number;
  purchaseCost?: number;
  recipeUnit?: string;
  /** @nullable */
  category?: string | null;
}

export interface RecipeIngredient {
  ingredientId: number;
  ingredientName: string;
  quantity: number;
  unit: string;
  purchaseCost: number;
  purchaseUnit: string;
  purchaseUnitSize: number;
  recipeUnit: string;
  recipeUnitCost: number;
  lineCost: number;
}

export interface RecipeIngredientInput {
  ingredientId: number;
  quantity: number;
  unit: string;
}

export type MethodBlockType =
  (typeof MethodBlockType)[keyof typeof MethodBlockType];

export const MethodBlockType = {
  header: "header",
  numbered: "numbered",
  text: "text",
  subinstruction: "subinstruction",
} as const;

export interface MethodBlock {
  type: MethodBlockType;
  content: string;
  order: number;
}

export interface RecipeCostSummary {
  totalIngredientCost: number;
  servings: number;
  costPerPortion: number;
  wastageCost: number;
  wastagePercent: number;
  totalCostWithWastage: number;
  costPerPortionWithWastage: number;
  foodCostPercent: number;
  recommendedSalePrice: number;
}

export interface Recipe {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  servings: number;
  wastagePercent: number;
  foodCostPercent: number;
  tags: string[];
  allergens: string[];
  method: MethodBlock[];
  ingredients: RecipeIngredient[];
  costSummary: RecipeCostSummary;
  /** @nullable */
  authorName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface RecipeSummary {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  servings: number;
  tags: string[];
  allergens: string[];
  costPerPortion: number;
  recommendedSalePrice: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateRecipeBody {
  title: string;
  /** @nullable */
  description?: string | null;
  servings: number;
  wastagePercent: number;
  foodCostPercent: number;
  tags: string[];
  allergens: string[];
  method: MethodBlock[];
  ingredients: RecipeIngredientInput[];
  /** @nullable */
  authorName?: string | null;
}

export interface UpdateRecipeBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  servings?: number;
  wastagePercent?: number;
  foodCostPercent?: number;
  tags?: string[];
  allergens?: string[];
  method?: MethodBlock[];
  ingredients?: RecipeIngredientInput[];
  /** @nullable */
  authorName?: string | null;
}

export interface RecipeStats {
  totalRecipes: number;
  totalIngredients: number;
  avgCostPerPortion: number;
  avgRecommendedSalePrice: number;
  recentRecipes: RecipeSummary[];
}

export type ListIngredientsParams = {
  /**
   * Search term to filter ingredients
   */
  search?: string;
};

export type ListRecipesParams = {
  /**
   * Search by title or description
   */
  search?: string;
  /**
   * Filter by tag
   */
  tag?: string;
};