# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Recipe Coster API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: ingredients
    description: Ingredient/product management
  - name: recipes
    description: Recipe management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /ingredients:
    get:
      operationId: listIngredients
      tags: [ingredients]
      summary: List all ingredients
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search term to filter ingredients
      responses:
        "200":
          description: List of ingredients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ingredient"
    post:
      operationId: createIngredient
      tags: [ingredients]
      summary: Create a new ingredient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIngredientBody"
      responses:
        "201":
          description: Created ingredient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"

  /ingredients/{id}:
    get:
      operationId: getIngredient
      tags: [ingredients]
      summary: Get ingredient by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ingredient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateIngredient
      tags: [ingredients]
      summary: Update an ingredient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIngredientBody"
      responses:
        "200":
          description: Updated ingredient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteIngredient
      tags: [ingredients]
      summary: Delete an ingredient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /recipes:
    get:
      operationId: listRecipes
      tags: [recipes]
      summary: List all recipes
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by title or description
        - name: tag
          in: query
          schema:
            type: string
          description: Filter by tag
      responses:
        "200":
          description: List of recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecipeSummary"
    post:
      operationId: createRecipe
      tags: [recipes]
      summary: Create a new recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecipeBody"
      responses:
        "201":
          description: Created recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"

  /recipes/{id}:
    get:
      operationId: getRecipe
      tags: [recipes]
      summary: Get recipe by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateRecipe
      tags: [recipes]
      summary: Update a recipe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecipeBody"
      responses:
        "200":
          description: Updated recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRecipe
      tags: [recipes]
      summary: Delete a recipe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /recipes/stats/summary:
    get:
      operationId: getRecipeStats
      tags: [recipes]
      summary: Get recipe statistics summary
      responses:
        "200":
          description: Recipe stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Ingredient:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        supplier:
          type: ["string", "null"]
        purchaseUnit:
          type: string
          description: "The unit the ingredient is purchased in (e.g. kg, L, each)"
        purchaseUnitSize:
          type: number
          description: "The size of the purchase unit (e.g. 1 for 1kg, 5 for 5L)"
        purchaseCost:
          type: number
          description: "Cost of one purchase unit in dollars"
        recipeUnit:
          type: string
          description: "The unit used in recipes (e.g. g, ml, each)"
        recipeUnitCost:
          type: number
          description: "Cost per recipe unit (auto-calculated)"
        category:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - purchaseUnit
        - purchaseUnitSize
        - purchaseCost
        - recipeUnit
        - recipeUnitCost
        - createdAt
        - updatedAt

    CreateIngredientBody:
      type: object
      properties:
        name:
          type: string
        supplier:
          type: ["string", "null"]
        purchaseUnit:
          type: string
        purchaseUnitSize:
          type: number
        purchaseCost:
          type: number
        recipeUnit:
          type: string
        category:
          type: ["string", "null"]
      required:
        - name
        - purchaseUnit
        - purchaseUnitSize
        - purchaseCost
        - recipeUnit

    UpdateIngredientBody:
      type: object
      properties:
        name:
          type: string
        supplier:
          type: ["string", "null"]
        purchaseUnit:
          type: string
        purchaseUnitSize:
          type: number
        purchaseCost:
          type: number
        recipeUnit:
          type: string
        category:
          type: ["string", "null"]

    RecipeIngredient:
      type: object
      properties:
        ingredientId:
          type: integer
        ingredientName:
          type: string
        quantity:
          type: number
        unit:
          type: string
        purchaseCost:
          type: number
        purchaseUnit:
          type: string
        purchaseUnitSize:
          type: number
        recipeUnit:
          type: string
        recipeUnitCost:
          type: number
        lineCost:
          type: number
      required:
        - ingredientId
        - ingredientName
        - quantity
        - unit
        - purchaseCost
        - purchaseUnit
        - purchaseUnitSize
        - recipeUnit
        - recipeUnitCost
        - lineCost

    RecipeIngredientInput:
      type: object
      properties:
        ingredientId:
          type: integer
        quantity:
          type: number
        unit:
          type: string
      required:
        - ingredientId
        - quantity
        - unit

    MethodBlock:
      type: object
      properties:
        type:
          type: string
          enum: [header, numbered, text, subinstruction]
        content:
          type: string
        order:
          type: integer
      required:
        - type
        - content
        - order

    RecipeCostSummary:
      type: object
      properties:
        totalIngredientCost:
          type: number
        servings:
          type: integer
        costPerPortion:
          type: number
        wastageCost:
          type: number
        wastagePercent:
          type: number
        totalCostWithWastage:
          type: number
        costPerPortionWithWastage:
          type: number
        foodCostPercent:
          type: number
        recommendedSalePrice:
          type: number
      required:
        - totalIngredientCost
        - servings
        - costPerPortion
        - wastageCost
        - wastagePercent
        - totalCostWithWastage
        - costPerPortionWithWastage
        - foodCostPercent
        - recommendedSalePrice

    Recipe:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        servings:
          type: integer
        wastagePercent:
          type: number
        foodCostPercent:
          type: number
        tags:
          type: array
          items:
            type: string
        allergens:
          type: array
          items:
            type: string
        method:
          type: array
          items:
            $ref: "#/components/schemas/MethodBlock"
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredient"
        costSummary:
          $ref: "#/components/schemas/RecipeCostSummary"
        authorName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - servings
        - wastagePercent
        - foodCostPercent
        - tags
        - allergens
        - method
        - ingredients
        - costSummary
        - createdAt
        - updatedAt

    RecipeSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        servings:
          type: integer
        tags:
          type: array
          items:
            type: string
        allergens:
          type: array
          items:
            type: string
        costPerPortion:
          type: number
        recommendedSalePrice:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - servings
        - tags
        - allergens
        - costPerPortion
        - recommendedSalePrice
        - createdAt
        - updatedAt

    CreateRecipeBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        servings:
          type: integer
        wastagePercent:
          type: number
        foodCostPercent:
          type: number
        tags:
          type: array
          items:
            type: string
        allergens:
          type: array
          items:
            type: string
        method:
          type: array
          items:
            $ref: "#/components/schemas/MethodBlock"
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredientInput"
        authorName:
          type: ["string", "null"]
      required:
        - title
        - servings
        - wastagePercent
        - foodCostPercent
        - tags
        - allergens
        - method
        - ingredients

    UpdateRecipeBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        servings:
          type: integer
        wastagePercent:
          type: number
        foodCostPercent:
          type: number
        tags:
          type: array
          items:
            type: string
        allergens:
          type: array
          items:
            type: string
        method:
          type: array
          items:
            $ref: "#/components/schemas/MethodBlock"
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredientInput"
        authorName:
          type: ["string", "null"]

    RecipeStats:
      type: object
      properties:
        totalRecipes:
          type: integer
        totalIngredients:
          type: integer
        avgCostPerPortion:
          type: number
        avgRecommendedSalePrice:
          type: number
        recentRecipes:
          type: array
          items:
            $ref: "#/components/schemas/RecipeSummary"
      required:
        - totalRecipes
        - totalIngredients
        - avgCostPerPortion
        - avgRecommendedSalePrice
        - recentRecipes