# lib/api-zod/src/generated/types/createIngredientBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Recipe Coster API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateIngredientBody {
  name: string;
  /** @nullable */
  supplier?: string | null;
  purchaseUnit: string;
  purchaseUnitSize: number;
  purchaseCost: number;
  recipeUnit: string;
  /** @nullable */
  category?: string | null;
}