# lib/api-zod/src/generated/types/createRecipeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Recipe Coster API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MethodBlock } from "./methodBlock";
import type { RecipeIngredientInput } from "./recipeIngredientInput";

export interface CreateRecipeBody {
  title: string;
  /** @nullable */
  description?: string | null;
  servings: number;
  wastagePercent: number;
  foodCostPercent: number;
  tags: string[];
  allergens: string[];
  method: MethodBlock[];
  ingredients: RecipeIngredientInput[];
  /** @nullable */
  authorName?: string | null;
}