# lib/api-zod/src/generated/types/recipeIngredientInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Recipe Coster API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecipeIngredientInput {
  ingredientId: number;
  quantity: number;
  unit: string;
}